Move npm code #283
